# reflect: optimize VisibleFields a bit