The benchstat result:

% benchstat old.txt new.txt 
goos: darwin
goarch: arm64
pkg: reflect
                │   old.txt   │              new.txt               │
                │   sec/op    │   sec/op     vs base               │
VisibleFields-8   7.636µ ± 0%   7.561µ ± 0%  -0.98% (p=0.000 n=10)

                │   old.txt    │               new.txt               │
                │     B/op     │     B/op      vs base               │
VisibleFields-8   17.12Ki ± 0%   16.87Ki ± 0%  -1.46% (p=0.000 n=10)

                │  old.txt   │            new.txt             │
                │ allocs/op  │ allocs/op   vs base            │
VisibleFields-8   76.00 ± 0%   76.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal